Bot Configuration Refinement
Refined BotConfiguration to be static in nature and MoonBot class to correctly utilize this. The class utilize a private constructor to prevent instantiation and maps the property file keys and values to an immutable map for retrieval. If an error occurs it will return an empty properties map, if there is a value with a null or empty string it will send a warning to the user and will not attempt to map it. Added google commons library (Guava) for utility, it was used in this particular case for its Resources class. TODO: Look into options for defining fatal properties, as of now, all properties should be considered fatal if not found or there is an error reading them.